pyproject.toml
[project]
name = "aurite-txtai"
version = "0.1.0"
description = "AI-powered data processing and trend analysis"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.112.1",
    "uvicorn>=0.30.6",
    "sqlalchemy>=2.0.32",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "apscheduler>=3.10.4",
    "nltk>=3.8.1",
    "textblob>=0.18.0",
    "google-cloud-secret-manager>=2.20.2",
]

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict_optional = true 